Tienda/serializers/CategoriaSerializer.py
from rest_framework import serializers
from Tienda.models import Categoria

class CategoriaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Categoria
        fields = ['id', 'nombre']